pyproject.toml
[project]
name = "photodraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "defusedxml>=0.7.1",
    "exifread>=3.5.1",
    "pillow>=11.2.1",
    "pillow-avif-plugin>=1.5.2",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.11",
]
